+++ p2groep11.Net/Models/Domain/Parameter.cs
@@ -12,7 +12,7 @@ namespace p2groep11.Net.Models.Domain
 
 //De vraag zelf
 
-        public abstract int Execute(ClimateChart chart);
+        public abstract double Execute(ClimateChart chart);
         public abstract string[] GiveOptAnswers(ClimateChart chart);
         public abstract string GiveAnswer(ClimateChart chart);
     }
